tui: move the disk screens into tui/partitions.py
The closure of `partitions_screen` is fifty-one definitions with four names
crossing outward and none inward, so this boundary is one way; the earlier
attempt cut the same subject by hand and found edges both ways.
